feat: Support retrieval of Queues associated with the given Engine
