+++ tests/elis_api_client/test_engines.py
@@ -4,6 +4,7 @@ import pytest
 
 from rossum_api.domain_logic.resources import Resource
 from rossum_api.models.engine import Engine, EngineField
+from rossum_api.models.queue import Queue
 
 TEST_ENGINE_ID = 123
 
@@ -42,6 +43,17 @@ class TestEngine:
             assert engine == Engine(**dummy_engine)
 
         http_client.fetch_all.assert_called_with(Resource.Engine, (), ())
+
+    async def test_retrieve_engine_queue_stats(self, elis_client, dummy_queue, mock_generator):
+        client, http_client = elis_client
+        http_client.fetch_all.return_value = mock_generator(dummy_queue)
+
+        queues = client.retrieve_engine_queue_stats(TEST_ENGINE_ID)
+
+        async for queue in queues:
+            assert queue == Queue(**dummy_queue)
+
+        http_client.fetch_all.assert_called_with(Resource.Queue, engine=TEST_ENGINE_ID)
 
 
 @pytest.mark.asyncio
