+++ tests/elis_api_client/test_queues.py
@@ -11,62 +11,6 @@ from rossum_api.models.task import Task, TaskStatus, TaskType
 
 
 @pytest.fixture
-def dummy_queue():
-    return {
-        "id": 8198,
-        "name": "Received invoices",
-        "url": "https://elis.rossum.ai/api/v1/queues/8198",
-        "workspace": "https://elis.rossum.ai/api/v1/workspaces/7540",
-        "connector": None,
-        "webhooks": [],
-        "hooks": [],
-        "schema": "https://elis.rossum.ai/api/v1/schemas/31336",
-        "inbox": "https://elis.rossum.ai/api/v1/inboxes/1229",
-        "users": ["https://elis.rossum.ai/api/v1/users/10775"],
-        "session_timeout": "01:00:00",
-        "rir_url": "https://all.rir.rossum.ai",
-        "rir_params": None,
-        "dedicated_engine": None,
-        "generic_engine": "https://api.elis.develop.r8.lol/v1/generic_engines/1",
-        "counts": {
-            "importing": 0,
-            "split": 0,
-            "failed_import": 0,
-            "to_review": 2,
-            "reviewing": 0,
-            "confirmed": 0,
-            "exporting": 0,
-            "postponed": 0,
-            "failed_export": 0,
-            "exported": 0,
-            "deleted": 0,
-            "purged": 0,
-            "rejected": 0,
-        },
-        "default_score_threshold": 0.8,
-        "automation_enabled": False,
-        "automation_level": "never",
-        "locale": "en_US",
-        "metadata": {},
-        "use_confirmed_state": False,
-        "document_lifetime": "01:00:00",
-        "settings": {
-            "columns": [{"schema_id": "tags"}],
-            "hide_export_button": True,
-            "automation": {"automate_duplicates": True, "automate_suggested_edit": False},
-            "rejection_config": {"enabled": True},
-            "email_notifications": {
-                "recipient": {"email": "john.doe@company.com", "name": "John Doe"},
-                "unprocessable_attachments": False,
-                "email_with_no_attachments": True,
-                "postponed_annotations": False,
-                "deleted_annotations": False,
-            },
-        },
-    }
-
-
-@pytest.fixture
 def dummy_annotation():
     return {
         "id": 315777,
